Begin migration to GoalProtocol/DataPointProtocol (#466)
This is the first in a set of PRs which will be the most disruptive part
of the transition to using CoreData types throughout the app. It
introduces DataPointProtocol and GoalProtocol protocols which are
implemented by both the existing memory-based types and the new CoreData
types. It then starts to bring the two closer together in signature, and
update some call sites to use the new shared protocols.

I don't expect these types to exist long term, but they make it possible
to have the code compile and run in an intermediate state during the
transition.

Testing
Verified the code compiles

+++ BeeKit/Managers/GoalManager.swift
@@ -256,7 +256,7 @@ public actor GoalManager {
         let todayGoals = goals.values.map { (goal) in
             let shortSlug = goal.slug.prefix(20)
             let limsum = goal.limsum ?? ""
-            return ["deadline" : goal.deadline.intValue, "thumbUrl": goal.cacheBustingThumbUrl, "limSum": "\(shortSlug): \(limsum)", "slug": goal.slug, "hideDataEntry": goal.hideDataEntry()] as [String : Any]
+            return ["deadline" : goal.deadline, "thumbUrl": goal.cacheBustingThumbUrl, "limSum": "\(shortSlug): \(limsum)", "slug": goal.slug, "hideDataEntry": goal.hideDataEntry()] as [String : Any]
         }
         return Array(todayGoals.prefix(3)) as Array<Any>
     }
